add snippet about possible widget problems
a number of problems with using widgets have been reported in CR workshops
through the years, but as we now only use JupyterLab and the installation
of ipywidgets now automatically configures JupyterLab to enable them, future
problems are hopefully unlikely. Corner cases are described in the ipywidgets
installation instructions which are linked in this commit.

This commit thus fixes #106, fixes #91, fixes #62 and fixes #89