remove %matplotlib inline as it seems to break interactivity
From a comment in the notes:

"I had trouble getting matplotlib to behave interactively with widgets when %inline is invoked. Maybe remove that? "

Additionally:
"I think matplotlib has a quick way to invoke its widget with %widget as well"

Removing the `%matplotlib inline` line seems to have fixed "reactivity" in the plot when moving the slider.